backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlmodel>=0.0.14

# Gemini API mode
google-genai>=0.3.0

# Browser mode (Gemini via Chrome)
playwright>=1.40.0

# Ollama mode
requests>=2.31.0

# RAG / book ingestion
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-text-splitters>=0.0.1
chromadb>=0.4.0

# Poker engine
PyPokerEngine

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
